Add transit mode selector
